+++ apps/backoffice-v2/src/common/components/organisms/Sidebar/Sidebar.Group.tsx
@@ -1,0 +1,17 @@
+import * as React from 'react';
+import { ctw } from '@ballerine/ui';
+
+export const SidebarGroup = React.forwardRef<HTMLDivElement, React.ComponentProps<'div'>>(
+  ({ className, ...props }, ref) => {
+    return (
+      <div
+        ref={ref}
+        data-sidebar="group"
+        className={ctw('relative flex w-full min-w-0 flex-col p-2', className)}
+        {...props}
+      />
+    );
+  },
+);
+
+SidebarGroup.displayName = 'SidebarGroup';