+++ apps/backoffice-v2/src/domains/business-reports/hooks/queries/useBusinessReportsQuery/useBusinessReportsQuery.tsx
@@ -1,8 +1,8 @@
-import { TBusinessReportType } from '@/domains/business-reports/types';
 import { useIsAuthenticated } from '@/domains/auth/context/AuthProvider/hooks/useIsAuthenticated/useIsAuthenticated';
 import { useQuery } from '@tanstack/react-query';
 import { businessReportsQueryKey } from '@/domains/business-reports/query-keys';
 import { isString } from '@/common/utils/is-string/is-string';
+import { MerchantReportType } from '@/domains/business-reports/constants';
 
 export const useBusinessReportsQuery = ({
   reportType,
@@ -12,7 +12,7 @@ export const useBusinessReportsQuery = ({
   sortBy,
   sortDir,
 }: {
-  reportType: TBusinessReportType;
+  reportType: MerchantReportType;
   search: string;
   page: number;
   pageSize: number;