+++ apps/backoffice-v2/src/lib/blocks/variants/WebsiteMonitoringBlocks/hooks/useWebsiteMonitoringReportBlock/useWebsiteMonitoringReportBlock.tsx
@@ -1,17 +1,17 @@
 import { createBlocksTyped } from '@/lib/blocks/create-blocks-typed/create-blocks-typed';
 import React, { useMemo } from 'react';
 import { useCurrentCaseQuery } from '@/pages/Entity/hooks/useCurrentCaseQuery/useCurrentCaseQuery';
-import { useLatestBusinessReportQuery } from '@/domains/business-reports/hooks/queries/useLatestBusinessReportQuery/useLatestBusinessReportQuery';
 import { WebsiteMonitoringBusinessReportTab } from '@/lib/blocks/variants/WebsiteMonitoringBlocks/hooks/useWebsiteMonitoringReportBlock/WebsiteMonitoringBusinessReportTab';
+import { useBusinessReportByIdQuery } from '@/domains/business-reports/hooks/queries/useBusinessReportByIdQuery/useBusinessReportByIdQuery';
 
 export const useWebsiteMonitoringReportBlock = () => {
   const { data: workflow } = useCurrentCaseQuery();
-  const { data: businessReport } = useLatestBusinessReportQuery({
-    businessId: workflow?.context?.entity?.ballerineEntityId,
-    reportType: 'MERCHANT_REPORT_T1',
+  const { data: businessReport } = useBusinessReportByIdQuery({
+    id: workflow?.context?.pluginsOutput?.merchantMonitoring?.reportId ?? '',
   });
+
   const blocks = useMemo(() => {
-    if (!businessReport?.report?.data) {
+    if (!businessReport?.data) {
       return [];
     }