+++ apps/backoffice-v2/src/pages/Entity/components/Case/hooks/useCaseActionsLogic/useCaseActionsLogic.tsx
@@ -1,15 +1,16 @@
-import { useWorkflowByIdQuery } from '@/domains/workflows/hooks/queries/useWorkflowByIdQuery/useWorkflowByIdQuery';
 import { useCallback, useMemo } from 'react';
-import { useDebounce } from '../../../../../../common/hooks/useDebounce/useDebounce';
-import { useFilterId } from '../../../../../../common/hooks/useFilterId/useFilterId';
-import { createInitials } from '../../../../../../common/utils/create-initials/create-initials';
-import { useAuthenticatedUserQuery } from '../../../../../../domains/auth/hooks/queries/useAuthenticatedUserQuery/useAuthenticatedUserQuery';
-import { useUsersQuery } from '../../../../../../domains/users/hooks/queries/useUsersQuery/useUsersQuery';
-import { useAssignWorkflowMutation } from '../../../../../../domains/workflows/hooks/mutations/useAssignWorkflowMutation/useAssignWorkflowMutation';
-import { tagToBadgeData } from '../../consts';
-import { useCaseDecision } from '../useCaseDecision/useCaseDecision';
-import { useCaseState } from '../useCaseState/useCaseState';
+
 import { IUseActions } from './interfaces';
+import { tagToBadgeData } from '../../consts';
+import { useCaseState } from '../useCaseState/useCaseState';
+import { useCaseDecision } from '../useCaseDecision/useCaseDecision';
+import { useDebounce } from '@/common/hooks/useDebounce/useDebounce';
+import { useFilterId } from '@/common/hooks/useFilterId/useFilterId';
+import { createInitials } from '@/common/utils/create-initials/create-initials';
+import { useUsersQuery } from '@/domains/users/hooks/queries/useUsersQuery/useUsersQuery';
+import { useWorkflowByIdQuery } from '@/domains/workflows/hooks/queries/useWorkflowByIdQuery/useWorkflowByIdQuery';
+import { useAuthenticatedUserQuery } from '@/domains/auth/hooks/queries/useAuthenticatedUserQuery/useAuthenticatedUserQuery';
+import { useAssignWorkflowMutation } from '@/domains/workflows/hooks/mutations/useAssignWorkflowMutation/useAssignWorkflowMutation';
 
 export const useCaseActionsLogic = ({ workflowId, fullName }: IUseActions) => {
   const filterId = useFilterId();