+++ apps/kyb-app/src/components/layouts/AppShell/Navigation.tsx
@@ -17,9 +17,12 @@ export const Navigation = () => {
   const { customer } = useCustomer();
   const { exit, isExitAvailable } = useAppExit();
 
-  const currentPageNumber = payload?.collectionFlow?.config?.steps?.find(
-    step => step.stateName === currentPage?.stateName,
-  )?.orderNumber;
+  const currentPageNumber =
+    Number(
+      payload?.collectionFlow?.state?.steps?.findIndex(
+        step => step.stepName === currentPage?.stateName,
+      ),
+    ) + 1;
 
   const isFirstStep = currentPageNumber === 1;
   const isDisabled = state.isLoading;
@@ -40,15 +43,15 @@ export const Navigation = () => {
 
   return (
     <button
-      className={ctw('cursor-pointer select-none', {
+      className={ctw('flex cursor-pointer select-none flex-row flex-nowrap items-center', {
         'pointer-events-none opacity-50': isDisabled,
       })}
       aria-disabled={isDisabled}
       onClick={onPrevious}
       type={'button'}
     >
-      <ArrowLeft className="inline" />
-      <span className="pl-2 align-middle text-sm font-bold">
+      <ArrowLeft size={24} className="flex-shrink-0" />
+      <span className="flex flex-nowrap pl-2 align-middle text-sm font-bold">
         {isFirstStep && customer
           ? t('backToPortal', { companyName: customer.displayName })
           : t('back')}