+++ packages/ui/src/components/organisms/DynamicForm/components/RSJVInputAdaters/PhoneInputAdapter/PhoneInputAdapter.tsx
@@ -7,17 +7,19 @@ export const PhoneInputAdapter: RJSFInputAdapter = ({
   formData,
   disabled,
   testId,
+  uiSchema,
   onChange,
   onBlur,
 }) => {
+  const { defaultCountry = 'us' } = uiSchema || {};
+
   const handleBlur = useCallback(() => {
-    // @ts-ignore
-    onBlur && onBlur(id, formData);
+    onBlur?.(id as string, formData);
   }, [id, onBlur, formData]);
 
   return (
     <PhoneNumberInput
-      country="us"
+      country={defaultCountry}
       value={formData}
       disabled={disabled}
       enableSearch