+++ services/workflows-service/src/business-report/business-report.module.ts
@@ -3,7 +3,6 @@ import { PrismaModule } from '@/prisma/prisma.module';
 import { ProjectModule } from '@/project/project.module';
 import { BusinessReportService } from '@/business-report/business-report.service';
 import { BusinessReportControllerInternal } from '@/business-report/business-report.controller.internal';
-import { BusinessReportRepository } from '@/business-report/business-report.repository';
 import { HttpModule } from '@nestjs/axios';
 // eslint-disable-next-line import/no-cycle
 import { DataAnalyticsModule } from '@/data-analytics/data-analytics.module';
@@ -15,6 +14,7 @@ import { EndUserModule } from '@/end-user/end-user.module';
 // eslint-disable-next-line import/no-cycle
 import { BusinessModule } from '@/business/business.module';
 import { CustomerModule } from '@/customer/customer.module';
+import { MerchantMonitoringClient } from '@/business-report/merchant-monitoring-client';
 import { BusinessReportControllerExternal } from '@/business-report/business-report.controller.external';
 
 @Module({
@@ -30,7 +30,7 @@ import { BusinessReportControllerExternal } from '@/business-report/business-rep
     BusinessModule,
     CustomerModule,
   ],
-  providers: [BusinessReportRepository, BusinessReportService],
-  exports: [BusinessReportRepository, BusinessReportService],
+  providers: [BusinessReportService, MerchantMonitoringClient],
+  exports: [BusinessReportService, MerchantMonitoringClient],
 })
 export class BusinessReportModule {}