+++ packages/react-pdf-toolkit/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @ballerine/react-pdf-toolkit
 
+## 1.2.43
+
+### Patch Changes
+
+- Updated dependencies
+  - @ballerine/ui@0.5.43
+
+## 1.2.42
+
+### Patch Changes
+
+- Change
+- Updated dependencies
+  - @ballerine/config@1.1.23
+  - @ballerine/ui@0.5.42
+
 ## 1.2.41
 
 ### Patch Changes