+++ services/workflows-service/CHANGELOG.md
@@ -1,5 +1,48 @@
 # @ballerine/workflows-service
 
+## 0.7.66
+
+### Patch Changes
+
+- version bump
+- Updated dependencies
+  - @ballerine/common@0.9.49
+  - @ballerine/workflow-core@0.6.62
+  - @ballerine/workflow-node-sdk@0.6.62
+
+## 0.7.65
+
+### Patch Changes
+
+- bump
+
+## 0.7.64
+
+### Patch Changes
+
+- Updated dependencies
+  - @ballerine/workflow-core@0.6.61
+  - @ballerine/workflow-node-sdk@0.6.61
+
+## 0.7.63
+
+### Patch Changes
+
+- Change
+- Updated dependencies
+  - @ballerine/common@0.9.48
+  - @ballerine/workflow-core@0.6.60
+  - @ballerine/workflow-node-sdk@0.6.60
+
+## 0.7.62
+
+### Patch Changes
+
+- Updated dependencies
+  - @ballerine/common@0.9.47
+  - @ballerine/workflow-core@0.6.59
+  - @ballerine/workflow-node-sdk@0.6.59
+
 ## 0.7.61
 
 ### Patch Changes